+++ src/qemu_vm.ts
@@ -36,7 +36,7 @@ export abstract class Vm extends vm.Vm {
       this.hypervisorPath.toString(),
       '-daemonize',
       '-machine', `type=${this.configuration.machineType},accel=${accel}`,
-      '-cpu', this.configuration.cpu,
+      '-cpu', this.cpuFlagValue,
       '-smp', this.configuration.cpuCount.toString(),
       '-m', this.configuration.memory,
 
@@ -48,9 +48,9 @@ export abstract class Vm extends vm.Vm {
       // '-nographic',
 
       '-boot', 'strict=off',
-      // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-      '-bios', this.configuration.firmware!.toString()
-    ].concat(this.hardDriverFlags)
+      ...this.firmwareFlags,
+      ...this.hardDriverFlags
+    ]
   }
 
   protected abstract get hardDriverFlags(): string[]
@@ -76,6 +76,14 @@ export abstract class Vm extends vm.Vm {
     return ''
   }
 
+  protected get cpuidFlags(): string[] {
+    return []
+  }
+
+  protected get firmwareFlags(): string[] {
+    return ['-bios', this.configuration.firmware!.toString()]
+  }
+
   private get netdev(): string {
     return [
       'user',
@@ -85,6 +93,10 @@ export abstract class Vm extends vm.Vm {
     ]
       .filter(e => e !== '')
       .join(',')
+  }
+
+  private get cpuFlagValue(): string {
+    return [this.configuration.cpu, ...this.cpuidFlags].join(',')
   }
 }